+++ pages/_app.js
@@ -5,6 +5,11 @@ import "../styles/global.css"
 const queryClient = new QueryClient();
 
 export default function App({ Component, pageProps }) {
+
+    if (typeof window === "undefined") {
+        return null;
+    }
+
     return (
         <QueryClientProvider client={queryClient}>
             <BackendProvider>
