+++ pages/index.js
@@ -1,57 +1,18 @@
-import { useState, useEffect } from 'react'
-import Featured from '../components/home/featured'
-import Hero from '../components/home/hero'
-import NearlyFunded from '../components/home/nearly-funded'
-import Navbar from '../components/shared/navbar'
-import { useBackend } from '@/context/backend'
-import Footer from '@/components/shared/footer'
-import LandingPopup, { LOCALSTORAGE_KEY } from '@/components/home/landing-popup'
+import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
+import HomePage from './home'
 
-function HomePage() {
-    const { backend, login } = useBackend()
-    const [profile, setProfile] = useState({})
-    const [profiles, setProfiles] = useState([])
-    const [test, setTest] = useState('')
-    const [showWelcomeMessage, setWelcomeMessage] = useState(true);
-
-    useEffect(() => {
-        if (!backend) return
-        backend.getOwnId().then((id) => setTest(id))
-        backend.getMyProfile().then((profile) => setProfile(profile))
-        backend.searchProfiles('s').then((profiles) => setProfiles(profiles))
-    }, [backend])
-
-    useEffect(() => {
-        const dontShow = localStorage.getItem(LOCALSTORAGE_KEY);
-    
-        if (dontShow === "true") {
-          setWelcomeMessage(false);
-        }
-      }, [])
-
-    async function sayGreeting() {
-        if (!backend) return alert('You must log in for this feature')
-        const greeting = await backend.greet()
-        alert(greeting)
-    }
-
-    if (showWelcomeMessage) {
-        return (<LandingPopup onContinue={() => setWelcomeMessage(false)} />);
-      }
-
+export default function Index() {
     return (
-        <div className='w-full'>
-            <Navbar />
-
-            <Hero />
-
-            <Featured />
-
-            {/* <NearlyFunded /> */}
-
-            <Footer />
+        <div>
+        <Router>
+            <div>                
+                <Switch>
+                    <Route path="/">
+                        <HomePage />
+                    </Route>
+                </Switch>
+            </div>
+        </Router>
         </div>
-    )
+    );
 }
-
-export default HomePage
